Match menu factory helper 00386860

+++ src/Event/mt_evtMenu.c
@@ -742,7 +742,7 @@ u32 FUN_00386310(int param_1);
 u64 FUN_00386430(int param_1,short param_2,u32 param_3);
 u32 FUN_003865f0(int param_1);
 u32 FUN_00386780(int param_1,u32 param_2,u32 param_3,u32 param_4);
-u64 FUN_00386860(int param_1,u32 param_2,u32 param_3,u32 param_4);
+u32 FUN_00386860(int param_1,u32 param_2,u32 param_3,u32 param_4);
 u64 FUN_00386940(u32 param_1,int param_2,u32 param_3);
 u32 FUN_00386a00(int param_1,u32 param_2);
 u64 FUN_0038da10(u16 *param_1);
@@ -19962,7 +19962,7 @@ u32 FUN_00386780(int param_1,u32 param_2,u32 param_3,u32 param_4)
 // FUN_00386860 NONMATCHING
 
 
-u64 FUN_00386860(int param_1,u32 param_2,u32 param_3,u32 param_4)
+u32 FUN_00386860(int param_1,u32 param_2,u32 param_3,u32 param_4)
 
 
 
@@ -19970,7 +19970,7 @@ u64 FUN_00386860(int param_1,u32 param_2,u32 param_3,u32 param_4)
 
   u32 *puVar1;
 
-  u64 uVar2;
+  u32 uVar2;
